Move bufread/write features under futures namespace
