+++ src/futures/bufread/generic/encoder.rs
