+++ src/futures/write/generic/decoder.rs
