+++ src/futures/write/generic/encoder.rs
