+++ src/futures/write/generic/mod.rs
