+++ src/futures/write/macros/mod.rs
